🐫👣 oh HaPpi DAY! simplify to asimple primaries. using just lost and
stack egg should be able to make everything work, rather then the
convoluted bubble balling lungi and all of that noise. this is
simplistic and should work allot better.

Listen to Oh Happi Dae by B💜aM🙊bi🩷♣️ on #SoundCloud
https://on.soundcloud.com/nHbWHLuruGEaZjfF7

sten to Oh Happi Dae by B💜aM🙊bi🩷♣️ on #SoundCloud
https://on.soundcloud.com/nHbWHLuruGEaZjfF7#	modified:   test/list.js
