--- src/lynkt_lyst.js
@@ -1,188 +1,0 @@
-
-var air;
-
-class LynktLyst {
-
-  static listOpenChar = "("; // SURR -> listOpenChar
-  static listCloseChar = ")"; // OUND -> listCloseChar
-
-  static get air() { return air; }
-  // static set air(value) {
-  //   switch (Math.trunc(Math.random() * 3)) {
-  //     case 0: throw new Error("You'rE tryiNg to overwrite LynktLyst.air, soOo i've got 1 question for you. Do you feel lucky...🤠...PuNk?!?");
-  //     case 1: throw new Error("👋 TheSe arE noT the dRoiDs yoU are LooKing for.");
-  //     case 2: throw new Error("http://youtu.be/otCpCn0l4Wo");
-  //   }
-  // };
-
-  get x() { return false; }
-  get first() { return peek(this); }
-  get rest() { return pop(this); }
-  get next() { return peek(pop(this)); }
-  get last() { return pop(this) ?
-      pop(this).last : peek(this); }
-
-  // Blow a lynktLyst.
-  constructor(o, oo) {
-    this.o=o;
-    this.oo=oo;
-  }
-
-  static make(...elements) { // oo -> elements
-    var listHead; // ooo -> listHead
-    for (let o of elements) // oo -> elements
-      listHead = new LynktLyst(o, listHead); // ooo -> listHead
-    return listHead; // ooo -> listHead
-  }
-
-  static from(arrayLike, mapFn, thisArg) { // arryLike -> arrayLike
-    let array = Array.from(arrayLike, mapFn, thisArg); // arryLike -> arrayLike
-    return LynktLyst.make(...array);
-  }
-
-  static get(o,i) { return peek(skip(o,i)); }
-
-  static skip(o, count) {
-    if (count)
-      return skip(pop(o), --count);
-    return o;
-  }
-
-  static push(list, element) { // oo -> list, o -> element
-    return new LynktLyst(element, list); // o -> element, oo -> list
-  }
-
-  static peek(o) { return o && o.o; }
-  static pop(o) { return o && o.oo; }
-
-  static count(o) {
-    return reduce(o, (count) => {
-      return ++count;
-    }, 0);
-  }
-
-  static shift(o) {
-    return invert(pop(invert(o)));
-  }
-
-  static invert(o) {
-    if (o)
-      // oo (accumulator), o (currentElement)
-      return reduce(pop(o), (accumulator, currentElement) => {
-        return push(accumulator, currentElement);
-      }, new LynktLyst(peek(o)));
-  }
-
-  static conj(targetList, sourceList) { // o -> targetList, oo -> sourceList
-    // oo (accumulator), o (currentElement)
-    return sourceList.reduce(function(accumulator, currentElement) {
-      return accumulator.push(currentElement);
-    }, targetList); // o -> targetList
-  }
-
-  static toString(o) {
-    if (!o) return this.listOpenChar + this.listCloseChar; // SURR -> listOpenChar, OUND -> listCloseChar
-
-    let format = (o) => {
-      switch (typeof o) {
-        case "string":
-          return '"' + o + '"';
-        case "symbol":
-          return Symbol.keyFor(o);
-        default:
-          return o.toString();
-      }
-    }
-
-    // oo (accumulatedString), o (formattedElement)
-    let join = (accumulatedString, formattedElement) => {
-      return formattedElement + " " + accumulatedString;
-    }
-
-    return this.listOpenChar + // SURR -> listOpenChar
-      reduce(map(o, format), join)
-         + this.listCloseChar; // OUND -> listCloseChar
-  }
-
-  static toArray(o) {
-    // oo (array), o (currentElement)
-    return reduce(o, (array, currentElement) => {
-      array.push(currentElement); return array; }, []);
-  }
-
-  static map(o, fn) {
-    if (o)
-      return new LynktLyst(fn(o.o),
-        map(o.oo, fn));
-  }
-
-  static reduce(o, fn, memo) {
-    if (o) {
-      let oo = pop(o);
-      if (oo)
-        if (memo !==  undefined)
-          return reduce(oo, fn,
-            fn(memo, peek(o)))
-        else
-          return reduce(oo, fn, peek(o));
-      else if(memo !== undefined)
-        return fn(memo, peek(o));
-      else
-        return peek(o);
-    } else
-      return memo;
-  }
-
-  static each(o, fn) {
-    let oo = fn(peek(o));
-    if (pop(o))
-      return each(pop(o), fn);
-    return oo;
-  }
-
-  //// Members Only ¥ ////
-
-  count() { return count(this); }
-  conj(sourceList) { return conj(this, sourceList); } // oo -> sourceList
-  each(fn) { return each(this, fn); }
-  get(i) { return get(this, i); }
-  invert() { return invert(this); }
-  map(fn) { return map(this, fn); }
-  peek() { return peek(this); }
-  pop() { return pop(this); }
-  push(element) { return push(this, element); } // o -> element
-  reduce(fn, memo) {
-    return reduce(this, fn, memo); }
-  shift() { shift(this); }
-  skip(n) { skip(this, n) };
-  toString() { return toString(this); }
-  toArray() { return toArray(this); }
-
-  *[Symbol.iterator]() {
-    let currentNode = this;
-    // EmptiLyst (which is LynktLyst.air) has an accessor `get x() { return true; }`
-    // Regular LynktLyst nodes have `get x() { return false; }`
-    // So, iterate as long as the current node is not an EmptiLyst.
-    while (currentNode && !currentNode.x) {
-      yield currentNode.o;
-      currentNode = currentNode.oo;
-    }
-  }
-}
-
-const _toString = LynktLyst.toString;
-const toString = _toString.bind(LynktLyst);
-toString.toString = _toString;
-LynktLyst.toString = toString;
-
-const {count, conj, each, get, invert,
-  map, peek, pop, push, reduce, shift,
-  skip, toArray } = LynktLyst
-
-class EmptiLyst extends LynktLyst {
-  get x() { return true; }
-}
-
-air = new EmptiLyst()
-
-module.exports = LynktLyst;
